requirements.txt
feedparser
requests
pytz
msal
tqdm
pyyaml
microsoftgraph-python
bilix
bs4
colorlog